requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
python-dotenv
requests
